+++ scio-extra/src/main/scala/com/spotify/scio/extra/sparkey/instances/ShardedSparkeyReader.scala
@@ -35,10 +35,10 @@ import scala.jdk.CollectionConverters._
 class ShardedSparkeyReader(val sparkeys: Map[Short, SparkeyReader], val numShards: Short)
     extends SparkeyReader {
   def hashKey(arr: Array[Byte]): Short =
-    Math.floorMod(MurmurHash3.bytesHash(arr, 1), numShards).toShort
+    Math.floorMod(MurmurHash3.bytesHash(arr, 1), numShards.toInt).toShort
 
   def hashKey(str: String): Short =
-    Math.floorMod(MurmurHash3.stringHash(str, 1), numShards).toShort
+    Math.floorMod(MurmurHash3.stringHash(str, 1), numShards.toInt).toShort
 
   override def getAsString(key: String): String = {
     val hashed = hashKey(key)
